look for arbitration opportunity beta

+++ src/api/ox/params/input_params_type/input_params_type.rs
@@ -2,7 +2,7 @@
 /// QuoteInputParams.1 = (amount, dec)
 /// QuoteInputParams.2 = buy token
 /// QuoteInputParams.3 = taker address
-pub type QuoteInputParams = (Option<String>, Option<(u128, u8)>, Option<String>, Option<String>);
+pub type QuoteInputParams = (Option<String>, Option<(f64, u8)>, Option<String>, Option<String>);
 
 /// OrderBookInputParams.0 = base token
 /// OrderBookInputParams.1 = quote token
